Move Extensions to its own folder and remove much of unused array extensions
