+++ Gizmo/Classes/Common/Extensions/Collection+Extensions.swift
