+++ Gizmo/Classes/Common/Extensions/Data+Extensions.swift
